+++ CHANGES.rst
@@ -12,6 +12,8 @@ Changelog
 - Documentation improvements (#123)
 - kw arg readability (#115)
 - Replaced usage of args with script_args in ipwhois_cli.py
+- Minor optimization in whois.py and online/test_whois.py
+- Added coveralls integration and re-enabled online tests with Travis CI
 
 0.13.0 (2016-04-18)
 -------------------